Update functionality
make commandline (instead of internal hard coded file)

works for both bootstrap and nrt
